Refactored transports on the client side
